refactor: refactor some cogs
- Add Message class for sending message/errors/etc
- Move Python files to src folder
