+++ src/bot.py
@@ -27,7 +27,7 @@ async def unload(extension):
 
 
 # Load files
-for filename in os.listdir('./cogs'):
+for filename in os.listdir('src/cogs'):
     if filename.endswith('.py'):
         client.load_extension(f'cogs.{filename[:-3]}')
